+++ docs/CHANGELOG.md
@@ -31,6 +31,7 @@
 ### bug fixes
 - TextEditor, Page-Module: improve usage of desc_txt, desc_md...
 - SearchPage/AlbumPage: fix spinner
+- fixed string-replace
 
 ### breaking changes
 - bumped up deps